requirements.txt
requests
pytest-watch
numpy
